Ignore inline html
To create a line break in a paragraph in markdown you can either add a <br/>
in inline html or add two trailing spaces. mdl by default enforces no trailing
whitespace and no inline html, so one of these rules have to be disabled. I
like the inline <br/> better because it is more explicit and you can see it
without an editor that highlights whitespace.